Fixed bug with symbols parser
The symbols parser had a bug in which the final position was wrong.
Not only this would mess with the positions of all parsers that came
after, but it would fail to compose with `lexeme`, which depends on the
correct position being reported by the parser given as argument.

Added some manual test cases to make sure this bug doesn't creep up
again and some property tests using `stream_data`.